This merge has not been tested apart from building c/c++ with
--disable-bootstrap.

get_nonzero_bits_as_range() needs to be looked at.

From-SVN: r258769

+++ libgo/runtime/go-runtime-error.c
@@ -55,7 +55,7 @@ enum
   GO_NIL = 11
 };
 
-extern void __go_runtime_error () __attribute__ ((noreturn));
+extern void __go_runtime_error (int32) __attribute__ ((noreturn));
 
 void
 __go_runtime_error (int32 i)
